Note, Block refactoring & other fixes from Dropbox (#194)
https://github.com/workduck-io/mexit-monorepo/pull/194

* Block checkbox not updating fix

* Navigation from Search fixed

* Save and clear buffer on unmount instead of location change

* Half init layout fixed

* Refactor model and shortcut removed

* Full screen loader styled changed

* Title is not changing on switching snippet fixed

* Show error on Node rename with border

* Refactoring note in diff namespace with same name

* Block refactoring fixed

* Patch added

+++ apps/webapp/src/Editor/Actions/useEditorBlockSelection.tsx
@@ -9,15 +9,7 @@ import {
   TNodeEntry
 } from '@udecode/plate'
 
-import {
-  mog,
-  updateIds,
-  NodeEditorContent,
-  generateTempId,
-  BlockType,
-  BlockMetaDataType,
-  getBlockMetadata
-} from '@mexit/core'
+import { mog, updateIds, NodeEditorContent, generateTempId, BlockType, getBlockMetadata } from '@mexit/core'
 
 import { defaultContent } from '../../Data/baseData'
 import useBlockStore from '../../Stores/useBlockStore'
@@ -80,9 +72,9 @@ export const useEditorBlockSelection = () => {
       insertNodes(editor, { type: ELEMENT_PARAGRAPH, id: generateTempId(), children: [{ text: '' }] }, { at: [0] })
   }
 
-  const deleteSelectedBlock = (): any => {
+  const deleteSelectedBlock = (deleteBlock?: boolean): any => {
     const editorBlocks = getEditorBlocks()
-    deleteContentBlocks(editorBlocks)
+    if (deleteBlock) deleteContentBlocks(editorBlocks)
     return editorBlocks
   }
